Fix TCC device data parsing and add debug logging
- Fix IndoorHumidity type mismatch (TCC returns float, not int)
- Change IndoorHumidity from int to float64 in UIData and ZoneData
- Add comprehensive debug logging for device fetch operations
- Add X-Requested-With header for AJAX requests
- Log successful device data fetch with temperature details
- Check for error page redirects in device data responses

Fixes issue where device fetch was failing silently due to JSON
parsing error on humidity field. Device now successfully fetched:
temp=65.0°F, heat=67.0, cool=73.0, mode=off

+++ internal/tcc/types.go
@@ -20,7 +20,7 @@ type ZoneData struct {
 	CurrentTemp      float64 `json:"DispTemperature"`
 	HeatSetpoint     float64 `json:"HeatSetpoint"`
 	CoolSetpoint     float64 `json:"CoolSetpoint"`
-	IndoorHumidity   int     `json:"IndoorHumidity"`
+	IndoorHumidity   float64 `json:"IndoorHumidity"`
 	SystemSwitchPos  int     `json:"SystemSwitchPosition"`
 	EquipmentStatus  int     `json:"EquipmentOutputStatus"`
 	IsFanRunning     bool    `json:"IsFanRunning"`
@@ -67,8 +67,8 @@ type UIData struct {
 	DispTemperature        float64 `json:"DispTemperature"`
 	HeatSetpoint           float64 `json:"HeatSetpoint"`
 	CoolSetpoint           float64 `json:"CoolSetpoint"`
-	IndoorHumidity         int     `json:"IndoorHumidity"`
-	OutdoorHumidity        int     `json:"OutdoorHumidity"`
+	IndoorHumidity         float64 `json:"IndoorHumidity"`
+	OutdoorHumidity        float64 `json:"OutdoorHumidity"`
 	OutdoorTemperature     float64 `json:"OutdoorTemperature"`
 	SystemSwitchPosition   int     `json:"SystemSwitchPosition"`
 	EquipmentOutputStatus  int     `json:"EquipmentOutputStatus"`
